requirements.txt
selenium
pandas
numpy
sqlalchemy
matplotlib
seaborn
plotly
python-dotenv
tqdm